% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_li.R
\name{boot_li}
\alias{boot_li}
\title{Multi-annual model of infection and vaccination (version 2)}
\usage{
boot_li(dat, ages = 0:18, vac_strategy, stat = "median")
}
\arguments{
\item{dat}{data set}

\item{ages}{vector of ages}

\item{vac_strategy}{character string indicating vaccination strategy}

\item{stat}{statistic to be calculated for bootstrapping (mean or median)}
}
\description{
This function bootstraps the number of lifetime infections obtained from model simulations to obtain confidence intervals
}
\keyword{morevac}
