% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params_file.R
\name{create_params_file}
\alias{create_params_file}
\title{This function creates parameter input file}
\usage{
create_params_file(n_sim = 100, n_indiv = 30000, max_age = 80,
  vac_cutoff = 10, start_year = 1820, end_year = 2028,
  pandemic_beta = 0.4, epidemic_beta = 0.2, lhc_size = 1000,
  out_file = "parameter_values", seed = NULL)
}
\arguments{
\item{n_sim}{number of simulations}

\item{n_indiv}{number of individuals}

\item{max_age}{maximum age of individuals}

\item{vac_cutoff}{age at which vaccination stops}

\item{start_year}{start simulation year (YYYY format)}

\item{end_year}{end simulation year (YYYY format)}

\item{pandemic_beta}{value of pandemic force of infection}

\item{epidemic_beta}{value of epidemic force of infection}

\item{lhc_size}{number of latin hypercube parameter combinations}

\item{out_file}{name of output file}
}
\value{
tibble with parameter values and writes csv file to the working directory with parameter values
}
\description{
This function creates parameter input file
}
\keyword{morevac}
