% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_func.R
\name{drift_func}
\alias{drift_func}
\title{Drift generating function}
\usage{
drift_func(nyears = 200, rate = 1)
}
\arguments{
\item{nyears}{number of years to determine drift for}

\item{rate}{rate of exponential distribution to draw from}
}
\value{
distance matrix
}
\description{
This function determines the amount of drift each year.
}
\keyword{morevac}
