% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohort_ar.R
\name{get_cohort_ar}
\alias{get_cohort_ar}
\title{Multi-annual model of infection and vaccination (version 2)}
\usage{
get_cohort_ar(inf_history, vac_history, ages, years, year_index)
}
\arguments{
\item{inf_history}{matrix of infection histories for each person}

\item{vac_history}{matrix of vaccination history for each person}

\item{ages}{matrix of ages of each person for every year}

\item{years}{vector of years to run simulation over (YYYY format)}

\item{year_index}{vector of index of years to get attack rates for}
}
\value{
data frame of attack rates within the cohort by year
}
\description{
This function gets attack rates for a rolling cohort.
}
\keyword{morevac}
