% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohorts.R
\name{get_cohorts}
\alias{get_cohorts}
\title{Multi-annual model of infection and vaccination (version 2)}
\usage{
get_cohorts(
  inf_history,
  vac_history,
  ages,
  enrollment_start_year = 2000,
  ncohorts = 10,
  total_year_range,
  length_study = 18,
  write.file = FALSE,
  file = "test"
)
}
\arguments{
\item{inf_history}{sparse matrix of infection histories for each person}

\item{vac_history}{sparse matrix of vaccination history for each person}

\item{ages}{matrix of ages of each person for every year}

\item{enrollment_start_year}{year for cohort enrollment to start (YYYY format)}

\item{ncohorts}{number of cohorts to enroll}

\item{total_year_range}{vector of years to run simulations over}

\item{length_study}{number of years to follow each cohort}

\item{write.file}{logical. if true, the infection histories and vaccination
histories are written to the working directory}

\item{file}{character string used as the prefix for the output file if write.file = TRUE}
}
\value{
data frame of attack rates within the cohort by year
}
\description{
This function gets attack rates for a rolling cohort.
}
\keyword{morevac}
