% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lifetime_inf.R
\name{get_lifetime_inf}
\alias{get_lifetime_inf}
\title{Get number of lifetime infections by age}
\usage{
get_lifetime_inf(inf_history, ages = 0:79, maxage = 80)
}
\arguments{
\item{inf_history}{number of simulations to run}

\item{ages}{number of individuals to be simulated}

\item{maxage}{range of years to follow cohort}
}
\value{
matrix of number of lifetime infections for each age. Rownames indicate current age
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
