% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infect.R
\name{infect}
\alias{infect}
\title{Infection function}
\usage{
infect(susceptibility = 1, foi = 0.2)
}
\arguments{
\item{susceptibility}{Value of an individual's susceptibility.}

\item{foi}{Number between 0 and 1 indicating the force of infection.}
}
\value{
Indicator of infection (0 = not infected, 1 = infected).
}
\description{
This function initializes the population before running the model.
}
\examples{
infect()
}
\keyword{morevac}
