% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_pop.R, R/morevac_functions.R,
%   R/morevac_functions2.R
\name{initialize_pop}
\alias{initialize_pop}
\title{Initialize population}
\usage{
initialize_pop(nindiv = 1000, maxage = 80, years = 200)

initialize_pop(nindiv = 1000, maxage = 80, years = 200)

initialize_pop(nindiv = 1000, maxage = 80, years = 200)
}
\arguments{
\item{nindiv}{Number of individuals in the simulated population. Defaults to 1000.}

\item{maxage}{Maximum age of individuals after which they die and are removed from the population. Defaults to 80.}

\item{years}{Number of years to simulate over. Defaults to 200.}

\item{nindiv}{Number of individuals in the simulated population. Defaults to 1000.}

\item{timepoints}{Number of timepoints to simulate over. Defaults to 10.}

\item{nindiv}{Number of individuals in the simulated population. Defaults to 1000.}

\item{years}{Number of timepoints to simulate over. Defaults to 10.}
}
\value{
List containing initialized infection history matrix, vaccination history matrix, susceptibility matrix, attack rate vector, and incidence vector.

List containing initialized infection history matrix, vaccination history matrix, susceptibility matrix, attack rate vector, and incidence vector.

List containing initialized infection history matrix, vaccination history matrix, susceptibility matrix, attack rate vector, and incidence vector.
}
\description{
This function initializes the population before running the model.

This function initializes the population before running the model.

This function initializes the population before running the model.
}
\examples{
initialize_pop()    
initialize_pop()    
initialize_pop()    
}
\keyword{morevac}
