% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiannual1.R
\name{multiannual1}
\alias{multiannual1}
\title{Multi-annual model of infection and vaccination (version 1)}
\usage{
multiannual1(n = 100, years = 5, vac_threshold = 3,
  vac_coverage = 0.5, mybeta = 0.15, mydelta = 0.2, mygamma = 0.4,
  biannual = FALSE, same_indiv = FALSE)
}
\arguments{
\item{n}{number of individuals to be simulated}

\item{years}{number of years to run simulation over}

\item{vac_threshold}{age at which an individual may be vaccinated}

\item{vac_coverage}{vaccination coverage}

\item{mybeta}{annual force of infection}

\item{mydelta}{drift parameter}

\item{mygamma}{protective effect of vaccine}

\item{biannual}{logical. annual or biannual vaccination?}

\item{same_indiv}{logical. are the same individuals vaccinated each year?}
}
\value{
list with three elements: 1) a list of infection histories and attack rates,
        2) a plot of annual attack rates, and 3) a plot of annual attack rates by
        vaccination scenario
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
