% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiannual2.R
\name{multiannual2}
\alias{multiannual2}
\title{Multi-annual model of infection and vaccination (version 2)}
\usage{
multiannual2(n = 1000, years = 200, maxage = 80, start_year = 1820,
  start_vac_year = 2000, start_vac_age = 3, vac_coverage = 0.5,
  beta_pandemic = 0.4, beta_epidemic = 0.15, delta_x = 0.2,
  delta_v = 0.2, mygamma = 0.4, suscept_func_version = 1,
  biannual = FALSE, rho = 0, return_ar_only = 0)
}
\arguments{
\item{n}{number of individuals to be simulated}

\item{years}{number of years to run simulation over}

\item{maxage}{maximum age of an individual (removed from population after maxage)}

\item{start_year}{year to start simulation}

\item{start_vac_age}{age at which an individual may be vaccinated}

\item{vac_coverage}{vaccination coverage}

\item{beta_pandemic}{force of infection when simulation starts}

\item{beta_epidemic}{annual force of infection (after initialization of model)}

\item{delta_x}{drift parameter from natural infection}

\item{delta_v}{drift parameter from vaccination}

\item{mygamma}{protective effect of vaccine}

\item{suscept_func_version}{integer indicating which susceptibility version to use.
1 = either-or, 2 = multiplicative.}

\item{biannual}{logical. annual or biannual vaccination?}

\item{rho}{correlation of vaccination}

\item{return_ar_only}{0 = no, 1 = return annual attack rates, 2 = return attack rate by age}

\item{vac_start_year}{year that vaccination starts}
}
\value{
list with two elements: 1) a list of infection histories and attack rates and
        2) a plot of annual attack rates by vaccination scenario
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
