% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attack_rates.R
\name{plot_attack_rates}
\alias{plot_attack_rates}
\title{Plot multi-annual attack rates}
\usage{
plot_attack_rates(dat, mean_ar = NULL, vac_start_year = NULL,
  y_max = 0.5, by = NULL)
}
\arguments{
\item{by}{Name of variable you want to plot attack rates by.}

\item{data}{data frame of annual attack rates. Should have two colums: Year and Attack_Rate.
If by_age=TRUE, dat should have a third column: Age_Group.}
}
\value{
plot of multi-annual attack rates with vertical dashed line indicating year of start of
        vaccination and a horizontal dashed line indicating mean attack rate prior to start of vaccination.
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
