% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lifetime_infections.R
\name{plot_lifetime_infections}
\alias{plot_lifetime_infections}
\title{Plot multi-annual attack rates}
\usage{
plot_lifetime_infections(dat, by_vac = FALSE)
}
\arguments{
\item{dat}{data frame of annual attack rates. Should have two colums: Age and Life_Inf.
If by_vac=TRUE then there should be a third column: Vac_Strategy.}

\item{by_vac}{Logical. If true, lifetime infections will be plotted by vaccintion strategy.}
}
\value{
plot of lifetime infections by age.
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
