% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/postprocessing_for_sim_results_for_rolling_cohort.R
\name{postprocess_sim_results_for_rolling_cohort}
\alias{postprocess_sim_results_for_rolling_cohort}
\title{Multi-annual model of infection and vaccination (version 2)}
\usage{
postprocess_sim_results_for_rolling_cohort(
  sim_dat,
  n_sim = 100,
  total_year_range = 1918:2028,
  length_study = 19,
  write_file = FALSE,
  file = "test"
)
}
\arguments{
\item{sim_dat}{list of sparse matrices of simulations results (no vac)}

\item{n_sim}{number of simulations performed (must be the same for each sim case)}

\item{total_year_range}{vector of years simulations were run over (YYYY format)}

\item{length_study}{number of years to get cohorts from}

\item{write_file}{logical. If TRUE, writes infection and vaccination histories for cohorts to csv file}

\item{file}{character string indicating the name of the output file. Only used when write.file = TRUE.}
}
\value{
data frame of attack rates within the cohort by year
}
\description{
This function post processes simulation results for a rolling cohort.
}
\keyword{morevac}
