% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sim_output.R
\name{process_sim_output}
\alias{process_sim_output}
\title{Processes simulation output into a data frame}
\usage{
process_sim_output(sim_out, j, year_range, age_range)
}
\arguments{
\item{sim_out}{list output from simulations}

\item{j}{index of first list level}

\item{year_range}{range of years to follow cohort}

\item{age_range}{range of ages to follow cohort}
}
\value{
data frame with attack rate and lifetime infections by year, age, and vaccination strategy
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
