% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_sim_data.R
\name{read_in_sim_data}
\alias{read_in_sim_data}
\title{Read in simulation output}
\usage{
read_in_sim_data(tags)
}
\arguments{
\item{tags}{list filename tags (should be of length 3)}
}
\value{
list of attack rate and lifetime infections data for each vac scenario
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
