% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run simulations of multi-annual model of infection and vaccination}
\usage{
run_sim(sim = 100, nindiv = 1000, year_range, age_range,
  vaccov = 0.5, version = 1, biannual = FALSE, rho = 0,
  flag = "annual")
}
\arguments{
\item{sim}{number of simulations to run}

\item{nindiv}{number of individuals to be simulated}

\item{year_range}{range of years to follow cohort}

\item{age_range}{range of ages to follow cohort}

\item{vaccov}{vaccination coverage, should be between 0 and 1 (inclusive)}

\item{version}{which vaccination version to use: 1 = either-or, 2 = multiplicative}

\item{filename}{prefix of outputted figures}
}
\value{
two plots 1) plot of annual attack rates for the cohort and 2) a plot of number of lifetime infections by age
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
