% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_2.R
\name{run_sim_2}
\alias{run_sim_2}
\title{Run simulations of multiannual() model of infection and vaccination}
\usage{
run_sim_2(sim = 100, n = 10000, years = 1820:2019, max_age = 80,
  start_vac_year = 2000, vac_cov = c(rep(0.5, 80)), betas = c(0.4,
  rep(0.2, 199)), vac_protect = 0.7, suscept_version = 2,
  vac_strategy = 1, rho = 0.9, wane = 0, take = 1,
  epsilon = 0.03, drift_off = FALSE, file.out = FALSE, tag = "",
  seed = NULL)
}
\arguments{
\item{sim}{number of simulations to run}

\item{nindiv}{number of individuals to be simulated}

\item{year_range}{range of years to follow cohort}

\item{age_range}{range of ages to follow cohort}

\item{vaccov}{vaccination coverage, should be between 0 and 1 (inclusive)}

\item{filename}{prefix of outputted figures}

\item{version}{which vaccination version to use: 1 = either-or, 2 = multiplicative}
}
\value{
two plots 1) plot of annual attack rates for the cohort and 2) a plot of number of lifetime infections by age
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
