% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims_all.R
\name{run_sims_all}
\alias{run_sims_all}
\title{This function runs sims for each combination of parameter value from parameter input file creates}
\usage{
run_sims_all(params_file, out_file = "test")
}
\arguments{
\item{params_file}{character string indicating the name of the file with input parameter values}

\item{out_file}{character string indicating name of output file}
}
\value{
writes csv files to the working directory with infection and vaccination histories for every row of params_file
}
\description{
This function runs sims for each combination of parameter value from parameter input file creates
}
\keyword{morevac}
