% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims_clust.R
\name{run_sims_clust}
\alias{run_sims_clust}
\title{This function runs sims for each combination of parameter value from parameter input file creates}
\usage{
run_sims_clust(
  n_sim = 20,
  n_indiv = 10000,
  max_age = 80,
  start_year = 1820,
  end_year = 2028,
  pandemic_beta = 0.4,
  epidemic_beta = 0.2,
  wane = 1,
  take = 1,
  exposure_penalty = 0,
  vac_protect = 0.7,
  vac_cov = 0.5,
  rho = 0.9,
  vac_cutoff = 10,
  id,
  out_file = "sim"
)
}
\arguments{
\item{n_sim}{number of simulations}

\item{n_indiv}{number of individuals in population}

\item{max_age}{maximum age on individuals before they are removed from the population}

\item{start_year}{start year of simulation (YYYY format)}

\item{end_year}{end year of simulation (YYYY format)}

\item{pandemic_beta}{force of infection in pandemic year}

\item{epidemic_beta}{force of infection in epidemic year}

\item{wane}{amount of waning}

\item{take}{amount of take}

\item{exposure_penalty}{exposure penalty value}

\item{vac_protect}{vaccine related protection}

\item{vac_cov}{vaccine coverage}

\item{rho}{amount of correlation of individuals repeatedly vaccinated}

\item{vac_cutoff}{age at which vaccination ends}

\item{id}{simulation number identifier}

\item{out_file}{character string indicating name of output file}
}
\value{
writes csv files to the working directory with infection and vaccination histories for every row of params_file
}
\description{
This function runs sims for each combination of parameter value from parameter input file creates
}
\keyword{morevac}
