% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_raw_output.R
\name{summarise_raw_output}
\alias{summarise_raw_output}
\title{This function runs sims for each combination of parameter value from parameter input file creates}
\usage{
summarise_raw_output(dt_inf, dt_vac, id)
}
\arguments{
\item{dt_inf}{data frame of infection histories}

\item{dt_vac}{data frame of vaccination histories}

\item{id}{index of parameter values}
}
\value{
returns a list of 3 data frames:
        - mean number of infections by vaccination strategy
        - mean difference in number of infections by vaccination strategy
        - mean attack rates by age and vaccination strategy
}
\description{
This function runs sims for each combination of parameter value from parameter input file creates
}
\keyword{morevac}
