% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suscept_func.R
\name{suscept_func}
\alias{suscept_func}
\title{Susceptibility function}
\usage{
suscept_func(inf_history, vac_history = 999, gamma = 0.4,
  drift_x = 0.2, drift_v = 0.2, version = 1)
}
\arguments{
\item{inf_history}{Number years since last infection}

\item{vac_history}{Number of years since last vaccination}

\item{gamma}{Vaccine efficacy}

\item{drift_x}{Amount of drift from infection}

\item{drift_v}{Amount of drift from vaccination}

\item{version}{Which susceptibility function to use: 1 = either-or, 2 = multiplicative}
}
\value{
Numeric value of susceptibility
}
\description{
This function initializes the population before running the model.
}
\examples{
suscept_func()
}
\keyword{morevac}
