% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_dist_func.R
\name{vac_dist_func}
\alias{vac_dist_func}
\title{Calculate vaccine distance from current strain based on last vaccine update}
\usage{
vac_dist_func(years = 200, drift, vac_protect = 0.7, begin_vac = 181)
}
\arguments{
\item{years}{number of years}

\item{drift}{vector of drift values}

\item{vac_protect}{protection from vaccine}
}
\value{
vector of scaled drift values for each year
}
\description{
This function determines the amount of drift each year.
}
\keyword{morevac}
