% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccinate.R
\name{vaccinate}
\alias{vaccinate}
\title{Vaccination function}
\usage{
vaccinate(year = 1, vc = 0.5, vac_strategy = FALSE)
}
\arguments{
\item{year}{Year counter.}

\item{vc}{Vaccine coverage.}

\item{vac_strategy}{Logical. FALSE = annual, TRUE = biannual.}
}
\value{
Indicator of infection (0 = not infected, 1 = infected).
}
\description{
This function initializes the population before running the model.
}
\examples{
vaccinate()
}
\keyword{morevac}
