% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morevac_functions.R
\name{initialize_pop}
\alias{initialize_pop}
\title{Initialize population}
\usage{
initialize_pop(nindiv = 1000, timepoints = 10)
}
\arguments{
\item{nindiv}{Number of individuals in the simulated population. Defaults to 1000.}

\item{timepoints}{Number of timepoints to simulate over. Defaults to 10.}
}
\value{
List containing initialized infection history matrix, vaccination history matrix, susceptibility matrix, attack rate vector, and incidence vector.
}
\description{
This function initializes the population before running the model.
}
\examples{
initialize_pop()    
}
\keyword{morevac}
