% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiannual.R
\name{multiannual}
\alias{multiannual}
\title{Multi-annual model of infection and vaccination}
\usage{
<<<<<<< HEAD
multiannual(n = 10000, years = 1820:2019, max_age = 80,
  start_vac_year = 2000, vac_coverage = c(rep(0.5, 80)),
  betas = c(0.4, rep(0.2, 199)), vac_protect = 0.7, vac_strategy = 1,
  rho = 0.9, wane = 0, take = 1, drift_rate = 1,
  drift_off = FALSE, epsilon = 0, seed = NULL)
=======
multiannual(
  n = 10000,
  years = 1820:2019,
  max_age = 80,
  start_vac_year = 2000,
  vac_coverage = c(rep(0.5, 80)),
  betas = c(0.4, rep(0.2, 199)),
  vac_protect = 0.7,
  vac_strategy = 1,
  rho = 0.9,
  wane = 0,
  take = 1,
  drift_rate = 1,
  drift_off = FALSE,
  epsilon = 0,
  seed = NULL
)
>>>>>>> 863033e (update files in man/)
}
\arguments{
\item{n}{number of individuals to be simulated}

\item{years}{vector of years to run simulation over (YYYY format)}

\item{max_age}{maximum age of an individual (removed from population after max_age)}

\item{vac_coverage}{vaccination coverage}

\item{betas}{vector of force of infection parameters for every year}

\item{vac_protect}{protective effect of vaccine}

\item{vac_strategy}{Integer value indicating frequency of vaccination (1 = annual, 2 = biannual, 3 =triannual,...)}

\item{rho}{correlation of vaccination}

\item{wane}{amount of protection of vaccine due to waning (0, 1) (inclusive)}

\item{take}{percentage of vaccine take (0, 1) (inclusive)}

\item{seed}{set random seed}

\item{vac_start_year}{year that vaccination starts (YYYY)}

\item{start_vac_age}{age at which an individual may be vaccinated}

\item{stop_vac_age}{age at which vaccination stops}

\item{suscept_func_version}{integer indicating which susceptibility version to use.
1 = either-or, 2 = multiplicative.}
}
\value{
list with two elements: 1) a list of infection histories and attack rates and
        2) a plot of annual attack rates by vaccination scenario
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
