% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiannual2.R
\name{multiannual2}
\alias{multiannual2}
\title{Multi-annual model of infection and vaccination (version 2)}
\usage{
multiannual2(n = 1000, years = 200, maxage = 80, start_year = 1820,
  start_vac_year = 2000, start_vac_age = 2, stop_vac_age = 10,
  vac_coverage = 0.5, beta_pandemic = 0.4, beta_epidemic = 0.2,
  vac_protect = 0.7, suscept_func_version = 1, vac_strategy = 1,
  rho = 0, wane = 0, take = 1, seed = NULL)
}
\arguments{
\item{n}{number of individuals to be simulated}

\item{years}{number of years to run simulation over}

\item{maxage}{maximum age of an individual (removed from population after maxage)}

\item{start_year}{year to start simulation}

\item{start_vac_age}{age at which an individual may be vaccinated}

\item{stop_vac_age}{age at which vaccination stops}

\item{vac_coverage}{vaccination coverage}

\item{beta_pandemic}{force of infection when simulation starts}

\item{beta_epidemic}{annual force of infection (after initialization of model)}

\item{suscept_func_version}{integer indicating which susceptibility version to use.
1 = either-or, 2 = multiplicative.}

\item{vac_strategy}{Integer value indicating frequency of vaccination (1 = annual, 2 = biannual, 3 =triannual,...)}

\item{rho}{correlation of vaccination}

\item{wane}{amount of protection of vaccine due to waning (0, 1) (inclusive)}

\item{take}{percentage of vaccine take (0, 1) (inclusive)}

\item{seed}{set random seed}

\item{vac_start_year}{year that vaccination starts}

\item{mygamma}{protective effect of vaccine}
}
\value{
list with two elements: 1) a list of infection histories and attack rates and
        2) a plot of annual attack rates by vaccination scenario
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
