% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attack_rates.R
\name{plot_attack_rates}
\alias{plot_attack_rates}
\title{Plot multi-annual attack rates}
\usage{
plot_attack_rates(dat, by_vac = FALSE, by_age_group = FALSE,
  c_bands = FALSE, y_max = 0.5, no_legend = FALSE, legend_x = 0.95,
  legend_y = 0.95)
}
\arguments{
\item{dat}{data frame of annual attack rates. Should have two colums: Year and Attack_Rate.
If by_vac=TRUE, dat should have a third column: Vac_Strategy.}

\item{by_vac}{logical. if TRUE then plot by vaccination strategy.}

\item{c_bands}{logical. Plot confidence bands or not. Must have columns named Lower and Upper.}

\item{y_max}{maximum value for y-axis.}

\item{no_legend}{logical. if TRUE legend will be suppressed.}

\item{legend_x}{x position of legend.}

\item{legend_y}{y position of legend.}
}
\value{
plot of multi-annual attack rates with vertical dashed line indicating year of start of
        vaccination and a horizontal dashed line indicating mean attack rate prior to start of vaccination.
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
