% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/postprocessing_for_sim_results_for_rolling_cohort.R
\name{postprocess_sim_results_for_rolling_cohort}
\alias{postprocess_sim_results_for_rolling_cohort}
\title{Multi-annual model of infection and vaccination (version 2)}
\usage{
postprocess_sim_results_for_rolling_cohort(simdat, nsim = 100,
  total_year_range = 1820:2028, length_study = 19,
  write.file = FALSE, file = "test")
}
\arguments{
\item{nsim}{number of simulations performed (must be the same for each sim case)}

\item{total_year_range}{vector of years simulations were run over (YYYY format)}

\item{sim0}{array of simulations results (no vac)}

\item{sim1}{array of simulations results (annual vac)}

\item{sim2}{array of simulation results (biannual vac)}

\item{cohort_year_index}{vector of index of years to get attack rates for}
}
\value{
data frame of attack rates within the cohort by year
}
\description{
This function post processes simulation results for a rolling cohort.
}
\keyword{morevac}
