% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sim_output.R
\name{process_sim_output}
\alias{process_sim_output}
\title{Processes simulation output into a data frame}
\usage{
process_sim_output(
  sim_out,
  year_range = 1820:2019,
  age_range = 0:18,
  ve_out = FALSE
)
}
\arguments{
\item{sim_out}{list output from simulations}

\item{year_range}{range of years to follow cohort}

\item{age_range}{range of ages to follow cohort}
}
\value{
list of two data frames:
        1) data frame with attack rates averaged over simulations by year, age, and vaccination strategy
        2) data frame with lifetime infections by simulation, year, age, and vac strategy
}
\description{
This function initializes the population before running the model.
}
\keyword{morevac}
