% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_2.R
\name{run_sim_2}
\alias{run_sim_2}
\title{Run simulations of multiannual() model of infection and vaccination}
\usage{
run_sim_2(
  sim = 100,
  n = 30000,
  years = 1918:2028,
  max_age = 80,
  start_vac_year = 2000,
  vac_cov = c(rep(0.5, 80)),
  betas = c(0.4, rep(0.2, 110)),
  vac_protect = 0.7,
  vac_strategy = 1,
  rho = 0.9,
  wane = 0,
  take = 1,
  epsilon = 0.03,
  drift_off = FALSE,
  trim_age_mat = NULL,
  seed = NULL
)
}
\arguments{
\item{sim}{number of simulations to run}

\item{n}{number of individuals to be simulated}

\item{years}{vector of years in YYYY format}

\item{max_age}{maimum age of an individual before they are removed from the population}

\item{vac_cov}{vector of vaccination coverages for each age group (should be of length max_age)}

\item{betas}{vector of force of infetion for each year (should be same length as years)}

\item{vac_protect}{vaccine efficacy}

\item{vac_strategy}{integer indicating which vaccination strategy to use (0 = no vaccination, 1 = annual, 2 = biennial)}

\item{rho}{correlation of repeat vaccination in individuals (between 0 and 1 (inclusive))}

\item{wane}{amount of vaccine waning}

\item{take}{proportion of individuals who receive the vaccine and are protected}

\item{epsilon}{exposure penalty}

\item{drift_off}{logical. if TRUE there will be no antigenic drift}

\item{trim_age_mat}{year before which age matrix is subsetted. If the year 2000 is used, then age matrix will only contain ages for year 2000 and beyond.}

\item{seed}{seed for simulations}
}
\value{
list of infection histories, vaccination histories, and age matrices for each simulation
}
\description{
Run simulations of multiannual() model of infection and vaccination
}
\keyword{morevac}
