% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_update.R
\name{vaccine_update}
\alias{vaccine_update}
\title{Vaccine update function}
\usage{
vaccine_update(years_since_vac_update, antigenic_dist, d = 3)
}
\arguments{
\item{years_since_vac_update}{number of years since last vaccine update}

\item{antigenic_dist}{antigenic_distance between circulating strain and vaccine strain since last vaccine update}

\item{d}{euclidean distance threshold for vaccine update}
}
\value{
binary indicator of whether the vaccine will be updated (0 - no, 1 - yes)
}
\description{
This function determines the amount of drift each year.
}
\keyword{morevac}
